Add banner to campaigns
