+++ db/migrate/201810202124322_add_banner_image_to_campaigns.rb
@@ -1,0 +1,5 @@
+class AddBannerImageToCampaigns < ActiveRecord::Migration
+  def change
+    add_column :campaigns, :banner_image, :string
+  end
+end
